+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 0.2.0
+*   added 4 cryptocurrency pairs from Bybit exchange. The full list of available cryptocurrency pairs is:
+    BTC_USDT, ETH_USDT, SOL_USDT, WLKN_USDT. [Issue #4]
+
 ## 0.1.0
 *   added 6 more currency pairs from Moscow Exchange (MOEX). The full list of available currency pairs is:
     USD_RUB, EUR_RUB, CNY_RUB, TRY_RUB, EUR_USD, USD_KZT, RUB_KZT; [Issue #3]